+++ features/step_definitions/common_steps.rb
@@ -6,6 +6,11 @@ When /^(.*) within a cassette named "([^"]*)"$/ do |step_to_call, cassette_name|
   VCR.use_cassette(cassette_name) { step step_to_call }
 end
 
+When /^(.*) within a cassette named "([^"]*)" and match on (.*)$/ do |step_to_call, cassette_name, matchers|
+  matches = matchers.split(',').map { |m| m.gsub(/\s*/,'') }.map(&:to_sym)
+  VCR.use_cassette(cassette_name, :match_requests_on => matches) { step step_to_call }
+end
+
 Then /^the response should equal (.*)$/ do |expected_response|
   expected = case expected_response
   when /t/
@@ -30,6 +35,8 @@ Then /^the response should be (.*)$/ do |expected_response|
     true
   when /\d+/
     expected_response.to_i
+  when /empty/
+    []
   else
     expected_response
   end
@@ -42,6 +49,8 @@ Then /^the response type should be (.*)$/ do |type|
     @response.content_type.should =~ /application\/json/
   when 'html'
     @response.content_type.should =~ /text\/html/
+  when 'raw'
+    @response.content_type.should =~ /raw/
   end
 end
 
@@ -60,10 +69,22 @@ Then /^the response should contain (.*)$/ do |item|
   end
 end
 
+Then /^the response should contain:$/ do |string|
+  unescape(@response.body).should include(unescape(string))
+end
+
 Then /^the response (.*) link should contain:$/ do |type, table|
   table.hashes.each do |attributes|
     attributes.each do |key, val|
       @response.links.send(:"#{type}").should match /#{key}=#{val}/
     end
+  end
+end
+
+Then /^the response (.*) item (.*) should be (.*)$/ do |action, field, result|
+  if action == 'first'
+    @response.first.send(field).should eql result
+  else
+    @response.last.send(field).should eql result
   end
 end