+++ lib/github_api/authorizations.rb
@@ -7,6 +7,10 @@ module Github
       scopes
       add_scopes
       remove_scopes
+      note
+      note_url
+      client_id
+      client_secret
     ].freeze
 
     # Creates new OAuth Authorizations API
@@ -82,12 +86,12 @@ module Github
     def update(authorization_id, params={})
       _check_if_authenticated
       assert_presence_of authorization_id
-
       normalize! params
       filter! VALID_AUTH_PARAM_NAMES, params
 
       patch_request("/authorizations/#{authorization_id}", params)
     end
+    alias :edit :update
 
     # Delete an authorization
     #
@@ -97,9 +101,7 @@ module Github
     def delete(authorization_id, params={})
       _check_if_authenticated
       assert_presence_of authorization_id
-
       normalize! params
-      filter! VALID_AUTH_PARAM_NAMES, params
 
       delete_request("/authorizations/#{authorization_id}", params)
     end