+++ lib/github_api/issues.rb
@@ -25,10 +25,11 @@ module Github
       body
       resource
       mime_type
+      org
     ].freeze
 
     VALID_ISSUE_PARAM_VALUES = {
-      'filter'    => %w[ assigned created mentioned subscribed ],
+      'filter'    => %w[ assigned created mentioned subscribed all ],
       'state'     => %w[ open closed ],
       'sort'      => %w[ created updated comments ],
       'direction' => %w[ desc asc ],
@@ -67,17 +68,57 @@ module Github
 
     # List your issues
     #
+    # List all issues across all the authenticated user’s visible repositories
+    # including owned repositories, member repositories,
+    # and organization repositories.
+    #
+    # = Example
+    #  github = Github.new :oauth_token => '...'
+    #  github.issues.list
+    #
+    # List all issues across owned and member repositories for the
+    # authenticated user.
+    #
+    # = Example
+    #  github = Github.new :oauth_token => '...'
+    #  github.issues.list :user
+    #
+    # List all issues for a given organization for the authenticated user.
+    #
+    # = Example
+    #  github = Github.new :oauth_token => '...'
+    #  github.issues.list :org => 'org-name'
+    #
+    # List issues for a repository
+    #
+    # = Example
+    #  github = Github.new
+    #  github.issues.list :user => 'user-name', :repo => 'repo-name'
+    #
     # = Parameters
     # <tt>:filter</tt>
-    #  * <tt>assigned</tt>: Issues assigned to you (default)
-    #  * <tt>created</tt>: Issues assigned to you (default)
-    #  * <tt>mentioned</tt>: Issues assigned to you (default)
-    #  * <tt>subscribed</tt>: Issues assigned to you (default)
-    # <tt>:state</tt> - <tt>open</tt>, <tt>closed</tt>, default: <tt>open</tt>
-    # <tt>:labels</tt> - String list of comma separated Label names. Example: bug,ui,@high
-    # <tt>:sort</tt> - <tt>created</tt>, <tt>updated</tt>, <tt>comments</tt>, default: <tt>created</tt>
+    #  * <tt>assigned</tt>   Issues assigned to you (default)
+    #  * <tt>created</tt>    Issues created by you 
+    #  * <tt>mentioned</tt>  Issues mentioning you 
+    #  * <tt>subscribed</tt> Issues you've subscribed to updates for 
+    #  * <tt>all</tt>        All issues the user can see 
+    # <tt>:milestone</tt>
+    #  * Integer Milestone number
+    #  * <tt>none</tt> for Issues with no Milestone.
+    #  * <tt>*</tt>    for Issues with any Milestone
+    # <tt>:state</tt>  - <tt>open</tt>, <tt>closed</tt>, default: <tt>open</tt>
+    # <tt>:labels</tt> - String list of comma separated Label names.
+    #                   Example: bug,ui,@high
+    # <tt>:assignee</tt>
+    #  * String User login
+    #  * <tt>none</tt> for Issues with no assigned User.
+    #  * <tt>*</tt>    for Issues with any assigned User.
+    # <tt>:mentioned</tt> String User login
+    # <tt>:sort</tt> - <tt>created</tt>, <tt>updated</tt>, <tt>comments</tt>,
+    #                  default: <tt>created</tt>
     # <tt>:direction</tt> - <tt>asc</tt>, <tt>desc</tt>, default: <tt>desc</tt>
-    # <tt>:since</tt> - Optional string of a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ
+    # <tt>:since</tt>     - Optional string of a timestamp in ISO 8601
+    #                       format: YYYY-MM-DDTHH:MM:SSZ
     #
     # = Examples
     #  github = Github.new :oauth_token => '...'
@@ -88,40 +129,6 @@ module Github
     #    :sort   => 'comments',
     #    :direction => 'asc'
     #
-    def list(*args)
-      params = args.extract_options!
-      normalize! params
-      filter! VALID_ISSUE_PARAM_NAMES, params
-      # _merge_mime_type(:issue, params)
-      assert_valid_values(VALID_ISSUE_PARAM_VALUES, params)
-
-      response = get_request("/issues", params)
-      return response unless block_given?
-      response.each { |el| yield el }
-    end
-    alias :all :list
-
-    # List issues for a repository
-    #
-    # = Parameters
-    # <tt>:milestone</tt>
-    #  * Integer Milestone number
-    #  * <tt>none</tt> for Issues with no Milestone.
-    #  * <tt>*</tt> for Issues with any Milestone
-    # <tt>:state</tt> - <tt>open</tt>, <tt>closed</tt>, default: <tt>open</tt>
-    # <tt>:assignee</tt>
-    #  * String User login
-    #  * <tt>none</tt> for Issues with no assigned User.
-    #  * <tt>*</tt> for Issues with any assigned User.
-    # <tt>:mentioned</tt> String User login
-    # <tt>:labels</tt> - String list of comma separated Label names. Example: bug,ui,@high
-    # <tt>:sort</tt> - <tt>created</tt>, <tt>updated</tt>, <tt>comments</tt>, default: <tt>created</tt>
-    # <tt>:direction</tt> - <tt>asc</tt>, <tt>desc</tt>, default: <tt>desc</tt>
-    # <tt>:since</tt> - Optional string of a timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ
-    # <tt></tt>, default: <tt>due_date</tt>
-    # <tt>:direction</tt> - <tt>asc</tt>, <tt>desc</tt>, default: <tt>desc</tt>
-    #
-    # = Examples
     #  github = Github.new :user => 'user-name', :repo => 'repo-name'
     #  github.issues.list_repo :milestone => 1,
     #    :state  => 'open',
@@ -131,20 +138,42 @@ module Github
     #    :sort   => 'comments',
     #    :direction => 'asc'
     #
-    def list_repo(user_name, repo_name, params={})
-      set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo
-
+    def list(*args)
+      params = args.extract_options!
       normalize! params
-      filter! VALID_ISSUE_PARAM_NAMES, params
+      # filter! VALID_ISSUE_PARAM_NAMES, params
       # _merge_mime_type(:issue, params)
-      assert_valid_values(VALID_ISSUE_PARAM_VALUES, params)
+      # assert_valid_values(VALID_ISSUE_PARAM_VALUES, params)
 
-      response = get_request("/repos/#{user}/#{repo}/issues", params)
+      response = if (org = params.delete('org'))
+        get_request("/orgs/#{org}/issues", params)
+
+      elsif (user_name = params.delete('user')) &&
+            (repo_name = params.delete('repo'))
+
+        list_repo user_name, repo_name, params
+      elsif args.include? :user
+        get_request("/user/issues", params)
+      else
+        get_request("/issues", params)
+      end
       return response unless block_given?
       response.each { |el| yield el }
     end
-    alias :list_repository :list_repo
+    alias :all :list
+
+    # List issues for a repository
+    #
+    def list_repo(user_name, repo_name, params)
+      set :user => user_name, :repo => repo_name
+      assert_presence_of user, repo
+
+      filter! VALID_ISSUE_PARAM_NAMES, params
+      assert_valid_values(VALID_ISSUE_PARAM_VALUES, params)
+
+      get_request("/repos/#{user}/#{repo}/issues", params)
+    end
+    private :list_repo
 
     # Get a single issue
     #