+++ lib/github_api/issues/labels.rb
@@ -13,16 +13,34 @@ module Github
     # List all labels for a repository
     #
     # = Examples
-    #  github = Github.new :user => 'user-name', :repo => 'repo-name'
+    #  github = Github.new user: 'user-name', repo: 'repo-name'
     #  github.issues.labels.list
     #  github.issues.labels.list { |label| ... }
+    #
+    # Get labels for every issue in a milestone
+    #
+    # = Examples
+    #  github = Github.new
+    #  github.issues.labels.list 'user-name', 'repo-name', milestone_id: 'milestone-id'
+    #
+    # List labels on an issue
+    #
+    # = Examples
+    #  @github = Github.new
+    #  @github.issues.labels.list 'user-name', 'repo-name', issue_id: 'issue-id'
     #
     def list(user_name, repo_name, params={})
       set :user => user_name, :repo => repo_name
       assert_presence_of user, repo
       normalize! params
 
-      response = get_request("/repos/#{user}/#{repo}/labels", params)
+      response = if (milestone_id = params.delete('milestone_id'))
+        get_request("/repos/#{user}/#{repo}/milestones/#{milestone_id}/labels", params)
+      elsif (issue_id = params.delete('issue_id'))
+        get_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels", params)
+      else
+        get_request("/repos/#{user}/#{repo}/labels", params)
+      end
       return response unless block_given?
       response.each { |el| yield el }
     end
@@ -32,14 +50,14 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.issues.labels.find 'user-name', 'repo-name', 'label-id'
+    #  github.issues.labels.find 'user-name', 'repo-name', 'label-name'
     #
-    def get(user_name, repo_name, label_id, params={})
+    def get(user_name, repo_name, label_name, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, label_id
+      assert_presence_of user, repo, label_name
       normalize! params
 
-      get_request("/repos/#{user}/#{repo}/labels/#{label_id}", params)
+      get_request("/repos/#{user}/#{repo}/labels/#{label_name}", params)
     end
     alias :find :get
 
@@ -50,8 +68,8 @@ module Github
     #  <tt>:color</tt> - Required string - 6 character hex code, without leading #
     #
     # = Examples
-    #  github = Github.new :user => 'user-name', :repo => 'repo-name'
-    #  github.issues.labels.create :name => 'API', :color => 'FFFFFF'
+    #  github = Github.new user: 'user-name', repo: 'repo-name'
+    #  github.issues.labels.create name: 'API', color: 'FFFFFF'
     #
     def create(user_name, repo_name, params={})
       set :user => user_name, :repo => repo_name
@@ -72,18 +90,18 @@ module Github
     #
     # = Examples
     #  @github = Github.new
-    #  @github.issues.labels.update 'user-name', 'repo-name', 'label-id',
-    #    :name => 'API', :color => "FFFFFF"
+    #  @github.issues.labels.update 'user-name', 'repo-name', 'label-name',
+    #    name: 'API', color: "FFFFFF"
     #
-    def update(user_name, repo_name, label_id, params={})
+    def update(user_name, repo_name, label_name, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, label_id
+      assert_presence_of user, repo, label_name
 
       normalize! params
       filter! VALID_LABEL_INPUTS, params
       assert_required_keys(VALID_LABEL_INPUTS, params)
 
-      patch_request("/repos/#{user}/#{repo}/labels/#{label_id}", params)
+      patch_request("/repos/#{user}/#{repo}/labels/#{label_name}", params)
     end
     alias :edit :update
 
@@ -91,30 +109,16 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.issues.labels.delete 'user-name', 'repo-name', 'label-id'
+    #  github.issues.labels.delete 'user-name', 'repo-name', 'label-name'
     #
-    def delete(user_name, repo_name, label_id, params={})
+    def delete(user_name, repo_name, label_name, params={})
       set :user => user_name, :repo => repo_name
       assert_presence_of user, repo
 
-      assert_presence_of label_id
+      assert_presence_of label_name
       normalize! params
 
-      delete_request("/repos/#{user}/#{repo}/labels/#{label_id}", params)
-    end
-
-    # List labels on an issue
-    #
-    # = Examples
-    #  @github = Github.new
-    #  @github.issues.labels.issue 'user-name', 'repo-name', 'issue-id'
-    #
-    def issue(user_name, repo_name, issue_id, params={})
-      set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, issue_id
-      normalize! params
-
-      get_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels", params)
+      delete_request("/repos/#{user}/#{repo}/labels/#{label_name}", params)
     end
 
     # Add labels to an issue
@@ -139,26 +143,27 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.issues.labels.remove 'user-name', 'repo-name', 'issue-id', 'label-id'
+    #  github.issues.labels.remove 'user-name', 'repo-name', 'issue-id',
+    #    lable_name: 'label-name'
     #
     # Remove all labels from an issue
     # = Examples
     #  github = Github.new
     #  github.issues.labels.remove 'user-name', 'repo-name', 'issue-id'
     #
-    def remove(user_name, repo_name, issue_id, label_id=nil, params={})
+    def remove(user_name, repo_name, issue_id, params={})
       set :user => user_name, :repo => repo_name
       assert_presence_of user, repo, issue_id
       normalize! params
 
-      if label_id
-        delete_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels/#{label_id}", params)
+      if (label_name = params.delete('label_name'))
+        delete_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels/#{label_name}", params)
       else
         delete_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels", params)
       end
     end
 
-    # Replace all labels for an issue 
+    # Replace all labels for an issue
     #
     # Sending an empty array ([]) will remove all Labels from the Issue.
     #
@@ -175,22 +180,6 @@ module Github
       normalize! params
 
       put_request("/repos/#{user}/#{repo}/issues/#{issue_id}/labels", params)
-    end
-
-    # Get labels for every issue in a milestone
-    #
-    # = Examples
-    #  github = Github.new
-    #  github.issues.labels. 'user-name', 'repo-name', 'milestone-id'
-    #
-    def milestone(user_name, repo_name, milestone_id, params={})
-      set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, milestone_id
-      normalize! params
-
-      response = get_request("/repos/#{user}/#{repo}/milestones/#{milestone_id}/labels", params)
-      return response unless block_given?
-      response.each { |el| yield el }
     end
 
   end # Issues::Labels