+++ lib/github_api/pull_requests.rb
@@ -62,18 +62,17 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.pull_requests.get 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.get 'user-name', 'repo-name', 'number'
     #
-    #  @pull_reqs = Github::PullRequests.new
-    #  @pull_reqs.get 'user-name', 'repo-name', 'request-id'
+    #  pull_reqs = Github::PullRequests.new
+    #  pull_reqs.get 'user-name', 'repo-name', 'number'
     #
-    def get(user_name, repo_name, request_id, params={})
+    def get(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
+      assert_presence_of user, repo, number
       normalize! params
-      # _merge_mime_type(:pull_request, params)
 
-      get_request("/repos/#{user}/#{repo}/pulls/#{request_id}", params)
+      get_request("/repos/#{user}/#{repo}/pulls/#{number}", params)
     end
     alias :find :get
 
@@ -101,7 +100,7 @@ module Github
     #
     # alternatively
     #
-    #  @github.pull_requests.create 'user-name', 'repo-name',
+    #  github.pull_requests.create 'user-name', 'repo-name',
     #    "issue" => "5",
     #    "head" => "octocat:new-feature",
     #    "base" => "master"
@@ -111,7 +110,6 @@ module Github
       assert_presence_of user, repo
 
       normalize! params
-      # _merge_mime_type(:pull_request, params)
       filter! VALID_REQUEST_PARAM_NAMES, params
 
       post_request("/repos/#{user}/#{repo}/pulls", params)
@@ -126,37 +124,33 @@ module Github
     #
     # = Examples
     #  github = Github.new :oauth_token => '...'
-    #  github.pull_requests.update 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.update 'user-name', 'repo-name', 'number'
     #    "title" => "Amazing new title",
     #    "body" => "Update body",
     #    "state" => "open",
     #
-    def update(user_name, repo_name, request_id, params={})
+    def update(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
-
+      assert_presence_of user, repo, number
       normalize! params
       filter! VALID_REQUEST_PARAM_NAMES, params
-      # _merge_mime_type(:pull_request, params)
       assert_valid_values(VALID_REQUEST_PARAM_VALUES, params)
 
-      patch_request("/repos/#{user}/#{repo}/pulls/#{request_id}", params)
+      patch_request("/repos/#{user}/#{repo}/pulls/#{number}", params)
     end
 
     # List commits on a pull request
     #
     # = Examples
     #  github = Github.new
-    #  github.pull_requests.commits 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.commits 'user-name', 'repo-name', 'number'
     #
-    def commits(user_name, repo_name, request_id, params={})
+    def commits(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
-
+      assert_presence_of user, repo, number
       normalize! params
-      # _merge_mime_type(:pull_request, params)
 
-      response = get_request("/repos/#{user}/#{repo}/pulls/#{request_id}/commits", params)
+      response = get_request("/repos/#{user}/#{repo}/pulls/#{number}/commits", params)
       return response unless block_given?
       response.each { |el| yield el }
     end
@@ -165,15 +159,14 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.pull_requests.files 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.files 'user-name', 'repo-name', 'number'
     #
-    def files(user_name, repo_name, request_id, params={})
+    def files(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
+      assert_presence_of user, repo, number
       normalize! params
-      # _merge_mime_type(:pull_request, params)
 
-      response = get_request("/repos/#{user}/#{repo}/pulls/#{request_id}/files", params)
+      response = get_request("/repos/#{user}/#{repo}/pulls/#{number}/files", params)
       return response unless block_given?
       response.each { |el| yield el }
     end
@@ -182,16 +175,14 @@ module Github
     #
     # = Examples
     #  github = Github.new
-    #  github.pull_requests.merged? 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.merged? 'user-name', 'repo-name', 'number'
     #
-    def merged?(user_name, repo_name, request_id, params={})
+    def merged?(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
-
+      assert_presence_of user, repo, number
       normalize! params
-      # _merge_mime_type(:pull_request, params)
 
-      get_request("/repos/#{user}/#{repo}/pulls/#{request_id}/merge", params)
+      get_request("/repos/#{user}/#{repo}/pulls/#{number}/merge", params)
       true
     rescue Github::Error::NotFound
       false
@@ -200,21 +191,21 @@ module Github
     # Merge a pull request(Merge Button)
     #
     # = Inputs
-    #  <tt>:commit_message</tt> - Optional string - The message that will be used for the merge commit
+    #  <tt>:commit_message</tt> - Optional string -
+    #                           The message that will be used for the merge commit
     #
     # = Examples
     #  github = Github.new
-    #  github.pull_requests.merge 'user-name', 'repo-name', 'request-id'
+    #  github.pull_requests.merge 'user-name', 'repo-name', 'number'
     #
-    def merge(user_name, repo_name, request_id, params={})
+    def merge(user_name, repo_name, number, params={})
       set :user => user_name, :repo => repo_name
-      assert_presence_of user, repo, request_id
+      assert_presence_of user, repo, number
 
       normalize! params
-      # _merge_mime_type(:pull_request, params)
       filter! VALID_REQUEST_PARAM_NAMES, params
 
-      put_request("/repos/#{user}/#{repo}/pulls/#{request_id}/merge", params)
+      put_request("/repos/#{user}/#{repo}/pulls/#{number}/merge", params)
     end
 
   end # PullRequests